Migrate API v1 code into apiV1 directory, rename function to streamQueryV1, and namespace types.
