+++ src/apiV1/index.ts
@@ -1,0 +1,1 @@
+export { streamQueryV1 } from "./client";
