+++ src/apiV2/index.ts
@@ -1,0 +1,1 @@
+export { streamQueryV2 } from "./client";
