+++ src/common/constants.ts
@@ -1,2 +1,4 @@
+export const DEFAULT_ENDPOINT = "api.vectara.io";
+
 export const SNIPPET_START_TAG = "%START_SNIPPET%";
 export const SNIPPET_END_TAG = "%END_SNIPPET%";
