+++ src/index.ts
@@ -1,2 +1,11 @@
-export { streamQuery } from "./client";
-export * from "./types";
+// Creates namespace "ApiV1"
+import * as ApiV1 from "./apiV1/types";
+export { ApiV1 };
+
+export { streamQueryV1 } from "./apiV1/client";
+
+// Creates namespace "ApiV2"
+// import * as ApiV2 from "./apiV2/types";
+// export { ApiV2 };
+
+export { streamQueryV2 } from "./apiV2/client";
